Revert "ni me acuerdo que cambio realize"
This reverts commit 67869eccf631a6b4f5bad6c0228c151c0c2def8f.

+++ src/main/java/com/miportfolio/proyecto/repository/RPersona.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RPersona extends JpaRepository <Persona, Long> {
+public interface RPersona extends JpaRepository <Persona, Integer> {
     public List<Persona> findByEmailAndClave(String email, String clave);
 }
